ZDoom 1.22 (zdoom122-src.zip)
https://zdoom.org/files/zdoom/1.x/zdoom122-src.zip

+++ code/d_main.cpp
@@ -1082,7 +1082,7 @@ void D_DoomMain (void)
 
 	atterm (DObject::StaticShutdown);
 
-	rngseed = (unsigned long)time (NULL);
+	rngseed = (DWORD)time (NULL);
 
 	gamestate = GS_STARTUP;
 	M_FindResponseFile ();
