+++ code/m_random.cpp
@@ -37,7 +37,7 @@
 //
 rng_t rng;     // the random number state
 
-unsigned long rngseed = 1993;   // killough 3/26/98: The seed
+DWORD rngseed = 1993;   // killough 3/26/98: The seed
 
 int P_Random (pr_class_t pr_class)
 {
@@ -67,7 +67,7 @@ int P_Random (pr_class_t pr_class)
 void M_ClearRandom (void)
 {
 	int i;
-	unsigned long seed = rngseed*2+1;	// add 3/26/98: add rngseed
+	DWORD seed = rngseed*2+1;	// add 3/26/98: add rngseed
 	for (i = 0; i < NUMPRCLASS; i++)	// go through each pr_class and set
 		rng.seed[i] = seed *= 69069ul;	// each starting seed differently
 }
